:arrow_up: add: 데이터베이스와 연결하도록 의존성 추가
