--- src/project/project.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('project')
-export class ProjectController {}
