--- src/project/project.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ProjectController } from './project.controller';
-import { ProjectService } from './project.service';
-
-@Module({
-  controllers: [ProjectController],
-  providers: [ProjectService]
-})
-export class ProjectModule {}
