+++ src/projects/project.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { ProjectsController } from './projects.controller';
+import { ProjectsService } from './projects.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Project } from './entities/project.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Project])],
+  controllers: [ProjectsController],
+  providers: [ProjectsService],
+})
+export class ProjectModule {}
