+++ src/projects/projects.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { ProjectsService } from './projects.service';
+
+@Controller('project')
+export class ProjectsController {
+  constructor(private readonly projectsService: ProjectsService) {}
+}
