+++ src/projects/projects.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@nestjs/common';
+import { ProjectsRepository } from './projects.repository';
+
+@Injectable()
+export class ProjectsService {
+  constructor(private readonly projectsRepository: ProjectsRepository) {}
+}
