+++ src/users/entities/user.entity.ts
@@ -1,5 +1,5 @@
 import { CoreEntity } from '../../commons/entities/core.entity';
-import { Project } from '../../project/entities/project.entity';
+import { Project } from '../../projects/entities/project.entity';
 import { Column, Entity, OneToMany } from 'typeorm';
 import { SKILL_TYPE, SKILL_TYPE_SET } from '../types/skill.type';
 import { on } from 'events';
@@ -92,6 +92,6 @@ export class User extends CoreEntity {
   @Column({ name: 'likes', default: 0, comment: '좋아요수' })
   likes: number;
 
-  @OneToMany(() => Project, (project) => project.user , { onDelete: 'CASCADE' })
+  @OneToMany(() => Project, (project) => project.user, { onDelete: 'CASCADE' })
   project: Project[];
 }
